[plumbing] Generate `AUTHORS.md` file from git
The git commit history stores the exact same information as the the
`AUTHORS` file.  Instead of managing this file separately and
duplicating the information, we should generate it from the git commit
history.

The `./prepare-release.sh` script does this.  It is named more generally
because I intend to generate other files from information stored in git
as well.

Finally, remove the old `AUTHORS` file and add `AUTHORS.md` to the
`.gitignore` file, to make sure it doesn't get committed again.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>