[plumbing] Deprecate GNU-style ChangeLog
A GNU-style ChangeLog, by its very nature, duplicates the information in
the git commit history.  We hope that our git commit history is a much
more useful than a ChangeLog, so instead of maintaining the information
in two places, we deprecate the old ChangeLog and rename it to show
this.

We don't provide a generator to recreate a GNU-style ChangeLog, because
this information is only useful to developers, who we assume will have
access to the git repository.  Any user-visible changes should be marked
in the `NEWS`.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>