[plumbing] Generate `NEWS.md` from git
Because we use the annotation of git tags to track user-visible release
notes, the `NEWS` file is superfluous.  As such, it is removed from
version control, and its contents generated by `./prepare-release.sh`.

To make sure it isn't accidentally committed back into the repository,
it's been added to the `.gitignore` file as well.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>