[plumbing] Add GPLv3+ license
All commits from `c84a3f5` on are released under GPLv3+, so we should
include this in the repository.  All other commits are released under
GPLv2+, which is compatible, so we keep the GPLv2 around as well.  Note
that this makes doxymacs effectively GPLv3+ from now on.

For easy viewing on GitHub, we've modified both to be Markdown.  The GPL
allows verbatim copies to be made, and this is generally interpreted to
deal with the text of the license, not the format, so we are good.

License: VerbatimCopying
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>