[plumbing] Convert `README` to Markdown
We would like to move entirely to Markdown, for the benefit of users
looking at the repo on GitHub: see issue #4 for this.  This patch adds
these first adaptions, namely Markdown section formatting, as well as
adds a packaging badge showing which GNU/Linux distributions currently
package Doxymacs.

License: GPLv3+
Reviewed-by: Patrick M. Niedzielski <patrick@pniedzielski.net>