Bigger restructuring of the package
Tested locally using non-GNU ELPA to build the package.
Therefore, the lisp subdirectory needed to be eliminated.
Automake could not be used any longer, and since I did
not know how to maintain that, I just wrote a makefile
and put the automake-processed *.el files into the top
level directory and now treat these as source files.
It appears not to be too unclean to be used this way,
but I'm welcoming other ideas as well