ci: Add a workflow for Gitlint
This patch adds a GitHub Actions workflow to run gitlint on each commit
in a pull request to `master`.  The configuration we lint against is
already included in the repository, but additionally for code going into
`master`, we do not want cleanup commits having titles starting with
`fixup!`, `squash!`, or `amend!` or having the string `WIP` in them, so
we turn on these checks additionally.

Fixes: #23
License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>