ci: Fix pull request target branch
Our default branch is currently `master`, which we inherited from the
previous code owners.  I’d like to switch to `main` (see #22), but until
we are able to do that, CI should use the `master` as the default
branch.  This patch changes our GitHub Actions workflow to run on pull
requests that target the corrent default branch.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>