ci: Checkout all commits for GitLint
The GitLint CI job needs access to all commits from this branch back to
the target branch, but by default, the checkout action only fetches the
most recent commit.  This patch checks out the entire history from the
repository, rather than just the most recent commit.  Because this
repository is not too large, the shortcut of checking out all commits in
the repository does not have much detriment.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>