refactor: Rename vendored `xml-parse.el`
Doxymacs vendors `xml-parse.el`, which was independently written by John
Wiegley and distributed long before the days of real Emacs packaging.
However, it does not seem to live in any package manager, so without
finding an alternative to migrate to, we can’t do much about this.  In
order to avoid clobbering an existing `xml-parse.el` from the user, this
patch renames `xml-parse.el` to `doxymacs-xml-parse.el` and prefixes all
symbols in the file with `doxymacs-xml-parse--`.

It is worth exploring using an alternative that is built-in to Emacs,
like `nxml-parser.el` or the libxml2 plugin if it’s enabled (#16).  For
the moment, though, it’s better to get this package building cleanly on
modern Emacsen with the minimal changes needed.

Note that `xml-parse.el` is GPLv2+ itself, but any changes in this patch
are licensed under GPLv3+.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>