build: Use pkg-config to find libxml2
We currently vendor an Autoconf M4 file to look for libxml2 on the
system.  Nowadays, it is much better to use the standard pkg-config:
[this is even what libxml2 itself does in newer versions of the file
we’ve vendored
in](https://github.com/GNOME/libxml2/commit/d598d8af0913b6e3d4e61ffa62397a275b669dca).
This patch removes the vendored file and switches to using pkg-config
to find libxml2.  This introduces a new dependency on pkg-config, which
we need to document; but, this is available on modern development
systems, so it should not be a burdensome dependency.

Ideally, if we fail to find either libxml2 or pkg-config, we should
avoid compiling the external XML parser but still create the elisp file,
configured not to support any external parser.  For now, though, this
patch makes no change to the current behavior.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>