fix: Add `lexical-binding` directives
This patch adds a `lexical-binding` directive to each of the package’s
elisp files, disabling lexical binding.  This makes the current binding
behavior explicit and silences the following warnings when
byte-compiling:

    Warning: file has no ‘lexical-binding’ directive on its first line

This patch does not enable lexical binding.  We should verify that the
current code works with lexical binding (it likely does) before enabling
it.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>