refactor: Remove support for XEmacs
We have no good tests for XEmacs, and because we are targeting modern
GNU Emacs, it is hard to know whether XEmacs will still work.  This
patch removes some special cases for XEmacs, giving us cleaner code to
work with.  Earlier versions of doxymacs will continue to work on
XEmacs, and if it is easy enough to support recent versions of XEmacs
and if there is enough support, it can be readded later.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>