refactor: Remove default keybindings
The default keybindings use reserved key sequences (`C-c` followed by a
character), so we cannot release this as a package.  Instead, it should
be up to the user to bind these keybindings in whatever major mode they
are using Doxygen with.  This patch removes the minor mode’s keymap.
When we are fully able to be a package, we will document the same
keybindings as the recommended ones for users to configure themselves,
including an example with `use-package`.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>