fix: Fix argument count for `string-trim`
Modern Emacsen allow `string-trim` to take two optional arguments,
denoting regular expressions that match what to trim on the left and
right of the string.  Older Emacsen that we do support do not take these
optional arguments.

We call `string-trim` with the two additional arguments as the regular
expressions `" +"` while looking for the names of thrown exceptions
inside a function body.  Luckily for us, this looks a bug in our code:
we want to remove all whitespace around a thrown exception’s name, not
just space characters.  This patch removes the extra arguments,
restoring compatibility with earlier Emacsen and improving our thrown
exception support.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>