fix: Replace list `defcustom`s with correct types
This patch replaces `defcustoms` that use `list` with ones that use the
correct types they expect.  For the tempo templates, this looks like
replacing them with `sexp`, which looks like it is what was intended.
For `doxymacs-doxygen-dirs`, this looks like replacing it with a more
structured list type, which allows the Customize interface to correctly
show it.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>