docs: Remove copy of GPLv2
This patch removes the copy of the GPLv2, and renames the copy of the
GPLv3.  All of the code from version 1.8 and before, and some of the
newer code, is licensed under the GNU GPLv2, but all other new
modifications since 1.8 have been licensed under the GNU GPLv3, so the
project as a whole must be distributed under that license.  The commit
messages make it clear what code is licensed how.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>