ci: Reorder CI checks
This patch changes the order of our CI checks, so that the simpler ones
appear first.  This does not affect the order they are performed, since
all checks are performed in parallel, but it does affect the order they
appear in the results.  It is better for GitLint and the version
consistency check to appear earlier, because they are easiest to fix.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>