feat: Add `doxymacs-install-external-parser`
This patch introduces the ability to build and install the external
parser from Emacs, in the fashion of pdf-tools.  Users configure
`doxymacs-external-xml-parser-executable` to whatever path they want,
and then run the interactive function `doxymacs-install-external-parser`
to build and install the parser to that location, if it doesn’t already
exist.

License: GPLv3+
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>